Move to correct script
